refactor: Extract `WebAnalyticsFilter` to separate file
